Fix type tests and try to fix type errors Fixes #2
- Also: cleaned pyproject.toml from redundancies

About the ignore flags in usecase.py:

Tried a couple of techniques, including using isinstance(), but none worked, or at least not without changing substantally what we did and entering a rabbit hole of new errors.
Digging further in SO, the consensus seems to be that this kind of type check is discouraged in python.
Also, the assertion by data and not by type seems stronger and better overall.